:white_check_mark: Reorganizing permissioning tests
