+++ tests/intermediate-level/acceleration.ts
@@ -5,20 +5,22 @@ import {
 import { expect } from "chai";
 
 export function acceleration(framework) {
-  it("Check component delegation to accelerator", async () => {
-    const delegateComponent = await DelegateComponent({
-      payer: framework.provider.wallet.publicKey,
-      entity: framework.entity1Pda,
-      componentId: framework.exampleComponentPosition.programId,
-    });
+  describe("Acceleration", () => {
+    it("Check component delegation to accelerator", async () => {
+      const delegateComponent = await DelegateComponent({
+        payer: framework.provider.wallet.publicKey,
+        entity: framework.entity1Pda,
+        componentId: framework.exampleComponentPosition.programId,
+      });
 
-    await framework.provider.sendAndConfirm(delegateComponent.transaction, [], {
-      skipPreflight: true,
-      commitment: "confirmed",
+      await framework.provider.sendAndConfirm(delegateComponent.transaction, [], {
+        skipPreflight: true,
+        commitment: "confirmed",
+      });
+      const acc = await framework.provider.connection.getAccountInfo(
+        delegateComponent.componentPda,
+      );
+      expect(acc?.owner.toBase58()).to.equal(DELEGATION_PROGRAM_ID.toBase58());
     });
-    const acc = await framework.provider.connection.getAccountInfo(
-      delegateComponent.componentPda,
-    );
-    expect(acc?.owner.toBase58()).to.equal(DELEGATION_PROGRAM_ID.toBase58());
   });
 }
