+++ tests/intermediate-level/index.ts
@@ -7,19 +7,9 @@ import { acceleration } from "./acceleration";
 
 describe("Intermediate level API", () => {
   let framework: Framework = new Framework();
-  describe("World", () => {
-    world(framework);
-  });
-  describe("ECS", () => {
-    ecs(framework);
-  });
-  describe("Session", () => {
-    session(framework);
-  });
-  describe("Permissioning", () => {
-    permissioning(framework);
-  });
-  describe("Acceleration", () => {
-    acceleration(framework);
-  });
+  world(framework);
+  ecs(framework);
+  session(framework);
+  permissioning(framework);
+  acceleration(framework);
 });
