+++ tests/low-level/permissioning/index.ts
@@ -1,0 +1,10 @@
+import { component } from "./component";
+import { world } from "./world";
+
+export function permissioning(framework) {
+  describe("Permissioning", () => {
+    component(framework);
+    world(framework);
+  });
+}
+
